requirements.txt
viam-sdk
pyinstaller
google-api-python-client
Adafruit_ADS1x15